BACK/NEEBIEBACK/requirements.txt
Django==5.1.4
channels==4.0.0
channels_redis==4.2.1
uvicorn==0.22.0
redis==5.0.0
djangorestframework==3.15.0
requests~=2.32.3
openai~=1.56.2
scikit-learn~=1.5.2
pytz~=2024.2
asgiref~=3.8.1
decouple~=0.0.7
python-decouple~=3.8